Android client setup documentation
This is a first draft of Android Setup instructions. I expect that these will need some revising and that the formatting of this document will benefit from more tweaking.

We may also want to elaborate on the contents of the "advanced settings" menu ([screenshot](https://i.imgur.com/smsmdQF.png)).