Fixes for #491 - Naming in mobileconfig, powershell, and AWS EC2
Normalized the various cloud_server_name entries to just
algo_server_name. Integrated algo_server_name into the mobileconfig
template as well as powershell. If a default name is provided, the
original behavior of IP_subject_alt_name will be used to avoid conflicts.
Also updated the Name tag in the cloudformation template to utilize the
StackName (which is based on the algo_server_name).